Refactor facility serializers by moving PartnerFieldEntrySerializer to a new file. This change improves code organization and maintainability by separating the partner field entry schema from the facility index details serializer.
